apps/backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

fastapi
python-multipart
Pillow
requests
torch==2.9.1
torchvision==0.24.1
uvicorn[standard]

